fixed location models
